Testing out kmalloc
