+++ src/memory/vmalloc.h
@@ -1,22 +1,8 @@
 #ifndef VMALLOC_H
 #define VMALLOC_H
 
+#include "memory/memory.h"
 #include <stddef.h>
-#include <stdint.h>
-
-#define HEAP_START (void *)0xD0000000
-#define MAGIC 0xDEADBEEF
-
-typedef struct free_list {
-  size_t size;
-  struct free_list *next;
-} free_list_t;
-
-typedef struct block_header {
-  uint8_t flags; // WIP, ignore for now
-  size_t size;
-  uint32_t magic;
-} block_header_t;
 
 void vfree(void *ptr);
